views/track: Merge into "tracking"
